completa separación de atributos
